backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
firebase-admin==6.8.0
pydantic==2.11.7
python-dotenv==1.1.1
reportlab==4.4.3
